Break up event model classes into domain-specific category submodules
