+++ streamdeck/models/events/system.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from typing import Literal
+
+from streamdeck.models.events.base import EventBase
+
+
+class SystemDidWakeUp(EventBase):
+    event: Literal["systemDidWakeUp"]  # type: ignore[override]
